LPD-55487 Remove unnecessary pagination test
