+++ modules/test/playwright/tests/portal-workflow-metrics-web/main/config.ts
@@ -1,0 +1,9 @@
+/**
+ * SPDX-FileCopyrightText: (c) 2025 Liferay, Inc. https://liferay.com
+ * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
+ */
+
+export const config = {
+	name: 'portal-workflow-metrics-web.main',
+	testDir: 'tests/portal-workflow-metrics-web/main',
+};
