+++ modules/test/playwright/tests/portal-workflow-metrics-web/main/test.properties
@@ -1,0 +1,5 @@
+##
+## Testray
+##
+
+    testray.main.component.name=Workflow Metrics
