+++ test.properties
@@ -2698,6 +2698,7 @@
         object-web.main,\
         portal-workflow-kaleo-designer-web.main,\
         portal-workflow-kaleo-forms-web.main,\
+        portal-workflow-metrics-web.main,\
         portal-workflow-task-web.main
 
     test.batch.class.names.excludes[bpm]=\
